problems with code
I am having issues with this code, it seems to work until I enter the
cost of gas, but it did not like it.

I tried changing the ID on the variable and still have issues with it.

it was originally a float, and is now a Double (ln32). if anyone knows
why this is, feel free to let me know/commit