Practicing creating new classes
I have practiced creating classes in these files,  from my textbook.

The idea behind this is to create a small program that will write people
a bank account and will hold it and can then output it. As the textbook
goes through ch. 3 more will be added to this.

Soon to push the reamaining files up.