D-buged code
I found out that the variables needed to be doubles not floats.

After this, I have set it to just output the numbers at the end. I will
comment that out of the final or I will go ahead and delete it.

I need to still check the math, it does not seem right to me.