added file form class computer
This file has been added to the repo, and it is using while loops and it also if statements. The source is from the textbook exercise 4.9. I have added extra lines of code on my own.

I have added the if statements to prevent the user from entering grades that are above 100%.