Code for the D&D project
Bassic source code for tmy D&D project.
Goal: input any number of entities names and an initive roll, reorder
them in order of least to greatest, and return the order in smallest to
largest or largest to smallest.

curetly implamented: ask user for amount of entities, and then store the
entities in different arrays (2). THen it will print out the two arrays,
however not in corredct order.