added conditional
so I made sure that if the card holder goes  over their limit, they will
incur a charge of $15, and their card will be blocked.

If they do not go over their limit, then the console prints out the new
balance, and what they last paid.

I am going to go ahead and add an intrest rate here in the next if
statment.