Lot of changes, but finished
lns16-21{
Created a method quotient that will accept two numbers and divide by
them.

this is where our error is
}

ln28-48{

Big loop that will continue so long as the bollean variable is true,
and that is only set to true if the try scenario does not return an
error.

ln39-43{

This is the code that will try something else if there is an error in
the code form the return statement.

}

}